Warnings cleanup, prefer size_t over unsigned for accessing containers

+++ Distribution2D.cpp
@@ -14,7 +14,7 @@ Distribution2D::Distribution2D(const float* values, size_t nu, size_t nv)
     _marginalDistributionV = std::make_unique<Distribution1D>(&_marginalDensityVValues[0], _nu);
 }
 
-float Distribution2D::sample(float e1, float e2, int &u, int &v) const
+float Distribution2D::sample(float e1, float e2, size_t& u, size_t& v) const
 {
     float marginal_density_v = _marginalDistributionV->sample(e1, u);
     float conditional_density_u = _conditionalDistributionU[u]->sample(e2, v);
